services/ocr-service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
pytesseract
opencv-python-headless
pdf2image
